Add module to save data
