+++ data/music_albums.json
@@ -1,6 +1,6 @@
 [
   {
     "on_spotify": true,
-    "Publish_date": "2020-01-01"
+    "Publish_date": "2022-12-04"
   }
 ]
